Change path for images

+++ src/medium-2-markdown/writer.js
@@ -3,8 +3,10 @@ const path = require('path')
 
 const write = (outputDir, file, data) => {
   const filePath = path.join(outputDir, `${file}.md`)
-  fs.writeFileSync(filePath, data)
-  return path.basename(filePath)
+  if (!fs.existsSync(filePath)) {
+    fs.writeFileSync(filePath, data)
+    return path.basename(filePath)
+  }
 }
 
 module.exports = write
